IMU should work now (thx claude :D)

+++ types/quaternion.h
@@ -11,8 +11,8 @@
  * @brief A struct to represent a quaternion value
  */
 struct quaternion {
-    int32_t x;
-    int32_t y;
-    int32_t z;
-    int32_t w;
+    float x;
+    float y;
+    float z;
+    float w;
 };
